Hotel Revenue Cleaning Function
Using python, I created this function to take inputs from other users to assign the working directory, file name, and year the revenue file is encompassing. The function requires two files Counties.csv and headers.csv. The Counties.csv file is a list of Texas counties and their corresponding Location County ID. The headers file is a transposed copy and pastes into a CSV from the data dictionary for column headings. The purpose of the function is to clean the raw data set and sum the reported monthly or quarterly revenues by property. I am in the process of making a Gross Revenue Multiplier tool to value hotels in texas.